Rename Thymeleaf3 to Thymeleaf and drop the old Thymeleaf version
