+++ vertx-template-engines/pom.xml
@@ -22,7 +22,6 @@
     <module>vertx-web-templ-mvel</module>
     <module>vertx-web-templ-handlebars</module>
     <module>vertx-web-templ-thymeleaf</module>
-    <module>vertx-web-templ-thymeleaf3</module>
     <module>vertx-web-templ-freemarker</module>
     <module>vertx-web-templ-pebble</module>
   </modules>
