+++ vertx-template-engines/vertx-web-templ-thymeleaf/pom.xml
@@ -16,7 +16,7 @@
     <dependency>
       <groupId>org.thymeleaf</groupId>
       <artifactId>thymeleaf</artifactId>
-      <version>2.1.2.RELEASE</version>
+      <version>3.0.0.RELEASE</version>
     </dependency>
     <dependency>
       <groupId>io.vertx</groupId>
