+++ vertx-template-engines/vertx-web-templ-thymeleaf/src/main/resources/vertx-web/thymeleaf3_template_engine.rb
