+++ vertx-template-engines/vertx-web-templ-thymeleaf/src/test/filesystemtemplates/test-thymeleaf-template3.html
@@ -1,7 +1,7 @@
-<!DOCTYPE html SYSTEM "http://www.thymeleaf.org/dtd/xhtml1-strict-thymeleaf-4.dtd">
+<!doctype html>
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
 <head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+  <meta charset="utf-8">
 </head>
 <body>
 <p th:text="${context.get('foo')}"></p>
