+++ vertx-template-engines/vertx-web-templ-thymeleaf/src/test/java/io/vertx/ext/web/templ/ThymeleafTemplateTest.java
@@ -47,11 +47,10 @@ public class ThymeleafTemplateTest extends WebTestBase {
     });
     router.route().handler(TemplateHandler.create(engine, directoryName, "text/html"));
     String expected =
-      "<!DOCTYPE html PUBLIC \"-//W3C//DTD XHTML 1.0 Strict//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd\">\n" +
-        "\n" +
+      "<!doctype html>\n" +
         "<html xmlns=\"http://www.w3.org/1999/xhtml\">\n" +
         "<head>\n" +
-        "  <meta http-equiv=\"Content-Type\" content=\"text/html; charset=UTF-8\" />\n" +
+        "  <meta charset=\"utf-8\">\n" +
         "</head>\n" +
         "<body>\n" +
         "<p>badger</p>\n" +
@@ -91,6 +90,12 @@ public class ThymeleafTemplateTest extends WebTestBase {
   public void testGetThymeLeafTemplateEngine() throws Exception {
     ThymeleafTemplateEngine engine = ThymeleafTemplateEngine.create();
     assertNotNull(engine.getThymeleafTemplateEngine());
+  }
+
+  @Test
+  public void testFragmentedTemplates() throws Exception {
+    TemplateEngine engine = ThymeleafTemplateEngine.create();
+    testTemplateHandler(engine, "somedir", "test-thymeleaf-fragmented.html");
   }
 
 }
