+++ vertx-web/src/test/resources/somedir/test-thymeleaf-fragmented.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
 </head>
 <body>
-<th:block th:include="somedir/test-thymeleaf3-include.html :: example"></th:block>
+<th:block th:include="somedir/test-thymeleaf-include.html :: example"></th:block>
 <p th:text="${context.normalisedPath()}"></p>
 <p th:text="${context.request().getParam('param1')}"></p>
 <p th:text="${context.request().getParam('param2')}"></p>
