--- vertx-web/src/test/resources/somedir/test-thymeleaf3-template2.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="utf-8">
-</head>
-<body>
-<p th:text="${context.get('foo')}"></p>
-<p th:text="${context.get('bar')}"></p>
-<p th:text="${context.normalisedPath()}"></p>
-<p th:text="${context.request().getParam('param1')}"></p>
-<p th:text="${context.request().getParam('param2')}"></p>
-</body>
-</html>
